requirements.txt
flask
werkzeug
Pillow
torch
